Carregar Imagem feito no MainActivity.

+++ app/src/main/java/com/example/corretordegabarito/Aluno.java
@@ -12,7 +12,7 @@ public class Aluno extends Foto {
      * @param formato - Formato da foto do tipo aluno
      */
 
-    public Aluno (int largura,int altura,String nomearquivo,long tamanho,String caminhoarquivo,String formato)
+    public Aluno (int largura,int altura,String nomearquivo,int tamanho,String caminhoarquivo,String formato)
     {
         super(largura,altura,nomearquivo,tamanho,caminhoarquivo,formato);
     }
