+++ app/src/main/java/com/example/corretordegabarito/Branco.java
@@ -12,7 +12,7 @@ public class Branco extends Foto {
      * @param formato - Formato da foto do tipo branco
      */
 
-    public Branco (int largura,int altura,String nomearquivo,long tamanho,String caminhoarquivo,String formato)
+    public Branco (int largura,int altura,String nomearquivo,int tamanho,String caminhoarquivo,String formato)
     {
         super(largura,altura,nomearquivo,tamanho,caminhoarquivo,formato);
     }
